refactor(storage): streamline import statements and clean up comments in chord sheet operations
