+++ frontend/src/storage/stores/chord-sheets/operations/get-chord-sheet.ts
@@ -1,10 +1,6 @@
-/**
- * Retrieves a chord sheet from IndexedDB storage
- */
-
-import type { Song } from '@chordium/types';
-import type { StoredChordSheet } from '../../../types/chord-sheet';
-import executeReadTransaction from '../utils/transactions/read-transaction';
+import type { Song } from "@chordium/types";
+import type { StoredChordSheet } from "../../../types/chord-sheet";
+import { executeReadTransaction } from "../utils/transactions";
 
 /**
  * Gets a stored chord sheet by its unique path identifier
@@ -12,8 +8,10 @@ import executeReadTransaction from '../utils/transactions/read-transaction';
  * @returns StoredChordSheet if found, null otherwise
  * @throws {DatabaseOperationError} When storage access fails
  */
-export default async function getChordSheet(path: Song["path"]): Promise<StoredChordSheet | null> {
-  return executeReadTransaction<StoredChordSheet | undefined>((store) => 
+export default async function getChordSheet(
+  path: Song["path"]
+): Promise<StoredChordSheet | null> {
+  return executeReadTransaction<StoredChordSheet | undefined>((store) =>
     store.get(path)
-  ).then(result => result || null);
+  ).then((result) => result || null);
 }
